extract e2e to new file
